a bunch of tidy up
