+++ app/src/lib/ssh/verification.ts
@@ -1,5 +1,12 @@
-import { exec } from 'child_process'
+import { exec, spawn } from 'child_process'
 import { getSSHEnvironment } from './ssh-environment'
+import * as fs from 'fs'
+import * as os from 'os'
+import * as Path from 'path'
+
+import { mkdirIfNeeded } from '../file-system'
+
+const processExists = require('process-exists')
 
 type HostVerificationError = {
   host: string
@@ -55,4 +62,46 @@ export function isHostVerificationError(
 export function isPermissionError(stderr: string): boolean {
   const permissionDeniedRe = /.*: Permission denied \(publickey\)\./
   return permissionDeniedRe.test(stderr)
+}
+
+export async function scanAndWriteToKnownHostsFile(
+  host: string
+): Promise<void> {
+  const sshDir = Path.join(os.homedir(), '.ssh')
+  await mkdirIfNeeded(sshDir)
+
+  const command = 'ssh-keyscan'
+  const env = await getSSHEnvironment(command)
+
+  return new Promise<void>((resolve, reject) => {
+    const keyscan = spawn(command, [host], { shell: true, env })
+    const homeDir = os.homedir()
+    const knownHostsPath = Path.join(homeDir, '.ssh', 'known_hosts')
+
+    keyscan.stdout.pipe(fs.createWriteStream(knownHostsPath))
+
+    keyscan.on('error', err => {
+      log.warn(
+        'Unable to execute ssh-keyscan and append to known_hosts file',
+        err
+      )
+    })
+
+    keyscan.on('close', code => {
+      if (code !== 0) {
+        reject(
+          new Error(
+            `ssh-keyscan exited with code '${code}' while adding '${host}' which was not expected`
+          )
+        )
+        return
+      }
+      resolve()
+    })
+  })
+}
+
+export async function findSSHAgentProcess(): Promise<boolean> {
+  const found: boolean = await processExists('ssh-agent')
+  return found
 }
